修改命令行输出cells的测试代码 用默认数据格式

+++ test/cellUpdateTest.py
@@ -1,7 +1,7 @@
 from src.cells import Cells
 import torch
 
-data = torch.zeros(10, 10).type(torch.LongTensor)
+data = torch.zeros(10, 10)
 data[0][2] = 1
 data[1][2] = 1
 data[2][2] = 1
